updated unit derrived stats to include full stat types. updated other classes accordingly. stats now use linear stacking

+++ src/Core/Domain/Effects/Stats/IDerivedStatsModifierSink.cs
@@ -5,5 +5,5 @@ namespace Core.Domain.Effects.Stats;
 public interface IDerivedStatsModifierSink
 {
     void ConsiderFlat(StatType stat, int delta, EffectInstanceId effectId);
-    void ConsiderPercent(StatType stat, float percentAdd, EffectInstanceId effectId);
+    void ConsiderPercent(StatType stat, int percentAdd, EffectInstanceId effectId);
 }
